backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.8.0
motor==3.5.0
pyotp==2.9.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
langgraph==0.2.20
langchain-openai==0.1.20
openai==1.51.0
pydantic==2.9.2
graphviz==0.20.3
matplotlib==3.9.2
requests==2.32.3
python-multipart==0.0.9
httpx==0.27.0
